fix: [MGS-4309] adjusted tests

+++ Test/Integration/TokenSecretHelper.php
@@ -7,6 +7,11 @@ class TokenSecretHelper
     public function prepareTokenSecretForTests()
     {
         $cacheFilePath = BP . '/var/global/lazy_resize_secret';
+
+        if (!file_exists($cacheFilePath)) { //phpcs:ignore
+            return;
+        }
+
         file_put_contents($cacheFilePath, (string)\MageSuite\LazyResize\Helper\Configuration::DEFAULT_TOKEN_SECRET); //phpcs:ignore
     }
 }
